+++ rollup.config.js
@@ -1,17 +1,38 @@
 import pkg from './package.json';
+import typescript from '@rollup/plugin-typescript';
+import { terser } from 'rollup-plugin-terser';
 
-export default {
-    input: 'src/index.js',
-    output: [
-		{
-			format: 'es',
-			sourcemap: true,
-			file: pkg.module,
-		},
+let output = [{
+	format: 'es',
+	file: pkg.module,
+	sourcemap: true,
+}];
+
+if (process.env.BUILD === 'production') {
+	output = [ ...output, ...[
 		{
 			format: 'cjs',
-			sourcemap: true,
 			file: pkg.main,
+			sourcemap: true,
 		},
+
+		{
+			format: 'umd',
+			file: pkg.unpkg,
+			name: 'FocusTrap',
+			sourcemap: true,
+			plugins: [ terser() ],
+		},
+	]];
+}
+
+export default {
+	input: 'src/index.ts',
+	output: output,
+	plugins: [
+		typescript({
+			declaration: false,
+			declarationDir: null,
+		}),
 	],
 };